Split `PathLayout.swift` to appease the linter
